examples/Odonto/requirements.txt
haystack-ai
markitdown-haystack
markitdown[pdf]
ollama-haystack
pdfminer.six
psycopg2-binary